refactor(scrollblockmanager): update tests to follow new standards
